Linux port: COMMS is the 13th module to build (via _COMM.CPP unity)
libbob_comms.a = _COMM.CPP (Aggrgtor/Comms/Replay/Winmove). DirectPlay multiplayer
stubbed — compiles but non-functional (lobby never connects).

Cross-cutting:
- compat_types.h: DWORD and ULONG are now `unsigned long` (Win32 ABI), matching
  bob's ULong. Fixes the 8 `cannot bind ULong& to DWORD` mismatches and is the
  faithful type. No regression to the 12 prior module libs (rebuilt clean).
- objbase.h: declaration-form DEFINE_GUID made variadic — bob writes the 1-arg
  DEFINE_GUID(BOB_GUID); form (values commented; definition in GENERAL/STATIC.CPP)
  relying on MSVC empty-fill; GCC is strict.
- dplobby.h: minimal IDirectPlayLobby3 stub (GetConnectionSettings/
  SetConnectionSettings/ConnectEx/Release) + CLSID_DirectPlayLobby /
  IID_IDirectPlayLobby3A GUIDs (the empty compat stub was shadowing the real SDK
  dplobby.h, leaving IDirectPlayLobby3 incomplete).

Per-file (REPLAY/WINMOVE): `extern BAD_RV;` -> `extern int` (implicit-int);
implicit-int statics (`static first=true` etc.) -> `static int`; two
`for(ULong n=0;...) ... return n;` for-scope hoists.

Thirteen module libs: math, lib3d, aircraft, general, files, input, missman, 3d,
ai, movecode, model, hardware, comms.

+++ SRC/compat/compat_types.h
@@ -121,9 +121,9 @@ typedef char                CHAR, *PCHAR;
 typedef unsigned short      WORD, *PWORD, *LPWORD;
 typedef short               SHORT, *PSHORT;
 typedef unsigned short      USHORT, *PUSHORT;
-typedef uint32_t            DWORD, *PDWORD, *LPDWORD;
+typedef unsigned long       DWORD, *PDWORD, *LPDWORD;	/* Win32 ABI: DWORD is unsigned long (matches bob ULong) */
 typedef long int            LONG, *PLONG, *LPLONG;	/* Win32: LONG is long (matches cstring.h) */
-typedef uint32_t            ULONG, *PULONG;
+typedef unsigned long       ULONG, *PULONG;
 typedef int                 INT, *PINT, *LPINT;
 typedef unsigned int        UINT, *PUINT;
 typedef float               FLOAT, *PFLOAT;
